Centrality_determination

+++ utilities/collect_events.sh
@@ -1,9 +1,17 @@
 #!/usr/bin/env bash
 
-usage="$0 fromFolder toFolder"
+# Determine the directory where this script resides
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+
+usage="$0 fromFolder toFolder [--destructive]"
 
 fromFolder=$1
 toFolder=$2
+destructive=false
+
+if [ "$3" == "--destructive" ]; then
+    destructive=true
+fi
 
 if [ -z "$fromFolder" ]
 then
@@ -19,7 +27,13 @@ fi
 fromFolder=${fromFolder%"/"}
 toFolder=${toFolder%"/"}
 
-echo "collecting events from " $fromFolder " to " $toFolder
+if [ "$destructive" = true ]; then
+    echo "collecting events from " $fromFolder " to " $toFolder " (destructive: moving files)"
+    move_or_copy=mv
+else
+    echo "collecting events from " $fromFolder " to " $toFolder " (non-destructive: copying files)"
+    move_or_copy=cp
+fi
 
 folderName=`echo $fromFolder | rev | cut -d "/" -f 1 | rev`
 target_folder=${toFolder}/${folderName}
@@ -59,16 +73,27 @@ do
             fi
         fi
         if [ -e $urqmd_file ]; then
-            urqmdstatus=true
+            # Find the URQMD folder (remove .gz and possible file extension)
+            urqmd_folder_path=$(dirname $urqmd_file)
+            if [ -d "$urqmd_folder_path" ]; then
+                if [ -z "$(ls -A "$urqmd_folder_path")" ]; then
+                    echo "URQMD folder $urqmd_folder_path is empty. Skipping this event."
+                    urqmdstatus=false
+                else
+                    urqmdstatus=true
+                fi
+            else
+                urqmdstatus=true
+            fi
         fi
         if [ -a ${eventsPath}/${iev}/${spvn_folder_name}*${event_id}.h5 ]; then
             if [ "$hydrostatus" = true ]; then
-                mv ${eventsPath}/${iev}/${hydro_folder_name}*${event_id} $target_hydro_folder
+                $move_or_copy -r ${eventsPath}/${iev}/${hydro_folder_name}*${event_id} $target_hydro_folder
             fi
             if [ "$urqmdstatus" = true ]; then
-                mv ${eventsPath}/${iev}/${UrQMD_file_name}*${event_id}.gz $target_urqmd_folder
+                $move_or_copy ${eventsPath}/${iev}/${UrQMD_file_name}*${event_id}.gz $target_urqmd_folder
             fi
-            mv ${eventsPath}/${iev}/${spvn_folder_name}*${event_id}.h5 $target_spvn_folder
+            $move_or_copy ${eventsPath}/${iev}/${spvn_folder_name}*${event_id}.h5 $target_spvn_folder
             ((collected_eventNum++))
         fi
         ((total_eventNum++))
@@ -78,8 +103,10 @@ done
 echo "Collected events number: " $collected_eventNum " out of " $total_eventNum
 
 if [ -f ${target_folder}/${folderName}.h5 ]; then
-    mv ${target_folder}/${folderName}.h5 ${target_spvn_folder}
+    $move_or_copy ${target_folder}/${folderName}.h5 ${target_spvn_folder}
 fi
-./combine_multiple_hdf5.py ${target_spvn_folder}
-mv SPVN_RESULTS.h5 ${target_folder}/${folderName}.h5
-rm -fr $target_spvn_folder
+"${SCRIPT_DIR}/combine_multiple_hdf5.py" ${target_spvn_folder}
+$move_or_copy SPVN_RESULTS.h5 ${target_folder}/${folderName}.h5
+if [ "$destructive" = true ]; then
+    rm -fr $target_spvn_folder
+fi
